// src/transaction/services/praxis/config/praxis-config.type.ts
export type PraxisConfigType = {
  api_url?: string;
  application_key?: string;
  domain?: string;
  version?: string;
  returnUrl?: string;
  notificationUrl?: string;
  merchantIdCreditCard?: string;
  merchantSecretCreditCard?: string;
  merchantIdCrypto?: string;
  merchantSecretCrypto?: string;
  merchantIdEWallet?: string;
  merchantSecretEWallet?: string;
};